[HIB-188] Move hyp3_daemon into hyp3_process module
