+++ ec2/worker/Pipfile
@@ -1,0 +1,13 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+"f9e3511" = {file = "file:///home/william/repositories/hyp3/in-a-box/modules/hyp3_events"}
+"61ea59b" = {path = "./../../modules/hyp3_events"}
+
+[dev-packages]
+
+[requires]
+python_version = "3.6"
