+++ ec2/worker/requirements.dev.txt
@@ -1,0 +1,2 @@
+-i https://pypi.org/simple
+
