+++ modules/hyp3_process/src/hyp3_process/__init__.py
@@ -1,1 +1,3 @@
 from .hyp3_process import Process, EarthdataCredentials, HandlerFunction
+from . import hyp3_daemon
+from .hyp3_daemon import log
