+++ modules/hyp3_process/src/hyp3_process/hyp3_daemon/hyp3_daemon.py
@@ -20,9 +20,10 @@ from multiprocessing import Pipe
 import boto3
 
 from hyp3_events import EmailEvent
-from hyp3_logging import getLogger
-from hyp3_worker import HyP3Worker, WorkerStatus
-from services import SNSService, SQSJob, SQSService
+
+from .hyp3_logging import getLogger
+from .hyp3_worker import HyP3Worker, WorkerStatus
+from .services import SNSService, SQSJob, SQSService
 
 log = getLogger(__name__)
