+++ modules/hyp3_process/src/hyp3_process/hyp3_daemon/hyp3_worker.py
