+++ modules/hyp3_process/src/hyp3_process/hyp3_daemon/services/__init__.py
@@ -10,7 +10,7 @@ from hashlib import md5
 import boto3
 from hyp3_events import EmailEvent, StartEvent
 
-from hyp3_logging import getLogger
+from ..hyp3_logging import getLogger
 
 log = getLogger(__name__)
